requirements.txt
rethinkdb==2.4.9
Flask==2.3.2
Flask-Session==0.5.0
Flask-SocketIO==5.3.5
gunicorn==21.2.0
gevent==23.7.0
python-dotenv==1.0.0
